+++ apps/charterafrica/src/components/EcosystemList/useEntity.js
@@ -1,18 +1,9 @@
 import useSWR from "swr";
 
+import queryString from "@/charterafrica/utils/ecosystem/queryString";
 import fetchJson from "@/charterafrica/utils/fetchJson";
 
 const fetcher = (url) => fetchJson.get(url);
-
-function queryString(params) {
-  const query = new URLSearchParams();
-  Object.keys(params).forEach((key) => {
-    if (params[key] && params[key] !== "undefined")
-      query.append(key, params[key]);
-  });
-  const qString = query.toString();
-  return qString ? `?${qString}` : "";
-}
 
 function useEntity(params) {
   const searchParams = queryString(params);
