+++ apps/charterafrica/src/pages/api/v1/resources/[type].page.js
@@ -1,8 +1,41 @@
 import * as Sentry from "@sentry/nextjs";
 
-import collections from "@/charterafrica/lib/ecosystem/collections";
+import {
+  getServerSideOrgs,
+  getServerSideContributors,
+  getServerSideTools,
+} from "@/charterafrica/lib/data/local";
 import fetchDatasets from "@/charterafrica/lib/openAfrica";
 import { fetchDocuments } from "@/charterafrica/lib/sourceAfrica";
+
+const collectionMap = {
+  organisations: getServerSideOrgs,
+  contributors: getServerSideContributors,
+  tools: getServerSideTools,
+};
+
+async function collections(req, res) {
+  const { query: { collection, locale = "en", ...query } = {} } = req;
+  const collectionFunc = collectionMap[collection];
+  if (!collectionFunc) {
+    return res.status(400).json({ message: "UNKNOWN_COLLECTION", collection });
+  }
+  const breadcrumbs = [
+    {
+      url: `/resources/${collection}`,
+    },
+  ];
+  const page = {
+    slug: collection,
+    breadcrumbs,
+  };
+  const context = {
+    locale,
+    query,
+  };
+  const found = await collectionFunc(page, context);
+  return res.status(200).json(found);
+}
 
 async function datasets(req, res) {
   const {
